requirements.txt
pygame>=2.1.0
Pillow>=9.0.0
requests>=2.28.0
pyserial>=3.5
